+++ cmd/server/main.go
@@ -4,11 +4,12 @@ import (
 	"context"
 	"encoding/hex"
 	"errors"
-	"github.com/rookgm/gophkeeper/internal/auth"
-	"github.com/rookgm/gophkeeper/internal/http/handler"
-	"github.com/rookgm/gophkeeper/internal/http/middleware"
-	"github.com/rookgm/gophkeeper/internal/repository"
-	"github.com/rookgm/gophkeeper/internal/service"
+	"github.com/rookgm/gophkeeper/internal/server/auth"
+	"github.com/rookgm/gophkeeper/internal/server/http/handler"
+	"github.com/rookgm/gophkeeper/internal/server/http/middleware"
+	"github.com/rookgm/gophkeeper/internal/server/repository"
+	"github.com/rookgm/gophkeeper/internal/server/repository/postgres"
+	"github.com/rookgm/gophkeeper/internal/server/service"
 	"net/http"
 	"os"
 	"os/signal"
@@ -17,7 +18,6 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/rookgm/gophkeeper/config"
 	"github.com/rookgm/gophkeeper/internal/logger"
-	"github.com/rookgm/gophkeeper/internal/repository/postgres"
 	"go.uber.org/zap"
 )
