+++ internal/server/http/handler/mocks/service_auth.go
